requirements.txt
annotated-types==0.5.0
anyio==3.7.1
attrs==23.2.0
cached-property==1.5.2
certifi==2024.6.2
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
importlib-metadata==6.7.0
openai==1.34.0
outcome==1.3.0.post0
pydantic==2.5.3
pydantic_core==2.14.6
PySocks==1.7.1
python-dotenv==0.21.1
selenium==4.11.2
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.66.4
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==1.26.6
wsproto==1.2.0
zipp==3.15.0
